requirements.txt
boto3==1.38.3
requests==2.32.3
python-dotenv==1.1
